Move storgae unittests
