--- UNITTESTS/storage/blockdevice/MBRBlockDevice/unittest.cmake
@@ -1,19 +1,0 @@
-
-####################
-# UNIT TESTS
-####################
-
-set(unittest-includes ${unittest-includes}
-  .
-  ..
-)
-
-set(unittest-sources
-  ../storage/blockdevice/MBRBlockDevice.cpp
-  stubs/mbed_atomic_stub.c
-  stubs/mbed_assert_stub.cpp
-)
-
-set(unittest-test-sources
-  storage/blockdevice/MBRBlockDevice/test_MBRBlockDevice.cpp
-)
