+++ storage/blockdevice/tests/UNITTESTS/blockdevice/SlicingBlockDevice/unittest.cmake
@@ -1,0 +1,20 @@
+
+####################
+# UNIT TESTS
+####################
+
+set(unittest-includes ${unittest-includes}
+  .
+  ..
+)
+
+set(unittest-sources
+  ../storage/blockdevice/source/SlicingBlockDevice.cpp
+  ../storage/blockdevice/source/HeapBlockDevice.cpp
+  stubs/mbed_atomic_stub.c
+  stubs/mbed_assert_stub.cpp
+)
+
+set(unittest-test-sources
+  ${CMAKE_CURRENT_LIST_DIR}/moduletest.cpp
+)
